WIP: Add way to ban item from search
