+++ pyproject.toml
@@ -121,8 +121,8 @@ ignore-words-list = 'nevers'
 minversion = "7.0"
 addopts = [
     "-sv",
-    # "--sw", # stepwise: resume with next test after failure
-    # "--pdb", # drop into debugger on failure
+    "--sw", # stepwise: resume with next test after failure
+    "--pdb", # drop into debugger on failure
     "-p no:warnings",
     # "--no-teardown", # don't teardown the database after tests
     # "--quiet-spy",                  # don't show logging from spyglass
