+++ tests/lfp/conftest.py
@@ -1,124 +1,5 @@
-import numpy as np
 import pytest
 from pynwb import NWBHDF5IO
-
-
-@pytest.fixture(scope="session")
-def firfilters_table(common):
-    return common.FirFilterParameters()
-
-
-@pytest.fixture(scope="session")
-def electrodegroup_table(lfp):
-    return lfp.v1.LFPElectrodeGroup()
-
-
-@pytest.fixture(scope="session")
-def lfp_constants(common, mini_copy_name, mini_dict):
-    n_delay = 9
-    lfp_electrode_group_name = "test"
-    orig_list_name = "01_s1"
-    orig_valid_times = (
-        common.IntervalList
-        & mini_dict
-        & f"interval_list_name = '{orig_list_name}'"
-    ).fetch1("valid_times")
-    new_list_name = orig_list_name + f"_first{n_delay}"
-    new_list_key = {
-        "nwb_file_name": mini_copy_name,
-        "interval_list_name": new_list_name,
-        "valid_times": np.asarray(
-            [[orig_valid_times[0, 0], orig_valid_times[0, 0] + n_delay]]
-        ),
-    }
-
-    yield dict(
-        lfp_electrode_ids=[0],
-        lfp_electrode_group_name=lfp_electrode_group_name,
-        lfp_eg_key={
-            "nwb_file_name": mini_copy_name,
-            "lfp_electrode_group_name": lfp_electrode_group_name,
-        },
-        n_delay=n_delay,
-        orig_interval_list_name=orig_list_name,
-        orig_valid_times=orig_valid_times,
-        interval_list_name=new_list_name,
-        interval_key=new_list_key,
-        filter1_name="LFP 0-400 Hz",
-        filter_sampling_rate=30_000,
-        filter2_name="Theta 5-11 Hz",
-        lfp_band_electrode_ids=[0],  # assumes we've filtered these electrodes
-        lfp_band_sampling_rate=100,  # desired sampling rate
-    )
-
-
-@pytest.fixture(scope="session")
-def add_electrode_group(
-    firfilters_table,
-    electrodegroup_table,
-    mini_copy_name,
-    lfp_constants,
-):
-    firfilters_table.create_standard_filters()
-    group_name = lfp_constants.get("lfp_electrode_group_name")
-    electrodegroup_table.create_lfp_electrode_group(
-        nwb_file_name=mini_copy_name,
-        group_name=group_name,
-        electrode_list=np.array(lfp_constants.get("lfp_electrode_ids")),
-    )
-    assert len(
-        electrodegroup_table & {"lfp_electrode_group_name": group_name}
-    ), "Failed to add LFPElectrodeGroup."
-    yield
-
-
-@pytest.fixture(scope="session")
-def add_interval(common, lfp_constants):
-    common.IntervalList.insert1(
-        lfp_constants.get("interval_key"), skip_duplicates=True
-    )
-    yield lfp_constants.get("interval_list_name")
-
-
-@pytest.fixture(scope="session")
-def add_selection(
-    lfp, common, add_electrode_group, add_interval, lfp_constants
-):
-    lfp_s_key = {
-        **lfp_constants.get("lfp_eg_key"),
-        "target_interval_list_name": add_interval,
-        "filter_name": lfp_constants.get("filter1_name"),
-        "filter_sampling_rate": lfp_constants.get("filter_sampling_rate"),
-    }
-    lfp.v1.LFPSelection.insert1(lfp_s_key, skip_duplicates=True)
-    yield lfp_s_key
-
-
-@pytest.fixture(scope="session")
-def lfp_s_key(lfp_constants, mini_copy_name):
-    yield {
-        "nwb_file_name": mini_copy_name,
-        "lfp_electrode_group_name": lfp_constants.get(
-            "lfp_electrode_group_name"
-        ),
-        "target_interval_list_name": lfp_constants.get("interval_list_name"),
-    }
-
-
-@pytest.fixture(scope="session")
-def populate_lfp(lfp, add_selection, lfp_s_key):
-    lfp.v1.LFPV1().populate(add_selection)
-    yield {"merge_id": (lfp.LFPOutput.LFPV1() & lfp_s_key).fetch1("merge_id")}
-
-
-@pytest.fixture(scope="session")
-def lfp_merge_key(populate_lfp):
-    yield populate_lfp
-
-
-@pytest.fixture(scope="session")
-def lfp_v1_key(lfp, lfp_s_key):
-    yield (lfp.v1.LFPV1 & lfp_s_key).fetch1("KEY")
 
 
 @pytest.fixture(scope="module")
