+++ tests/utils/test_chains.py
@@ -9,10 +9,15 @@ def TableChain():
     return TableChain
 
 
+def full_to_camel(t):
+    return to_camel_case(t.split(".")[-1].strip("`"))
+
+
 def test_chains_repr(chains):
     """Test that the repr of a TableChains object is as expected."""
     repr_got = repr(chains)
-    repr_exp = "\n".join([str(c) for c in chains.chains])
+    chain_st = ",\n\t".join([str(c) for c in chains.chains]) + "\n"
+    repr_exp = f"TableChains(\n\t{chain_st})"
     assert repr_got == repr_exp, "Unexpected repr of TableChains object."
 
 
@@ -32,11 +37,13 @@ def test_invalid_chain(Nwbfile, pos_merge_tables, TableChain):
 def test_chain_str(chain):
     """Test that the str of a TableChain object is as expected."""
     chain = chain
-    parent = to_camel_case(chain.parent.table_name)
-    child = to_camel_case(chain.child.table_name)
 
     str_got = str(chain)
-    str_exp = parent + chain._link_symbol + child
+    str_exp = (
+        full_to_camel(chain.parent)
+        + chain._link_symbol
+        + full_to_camel(chain.child)
+    )
 
     assert str_got == str_exp, "Unexpected str of TableChain object."
 
@@ -44,7 +51,9 @@ def test_chain_str(chain):
 def test_chain_repr(chain):
     """Test that the repr of a TableChain object is as expected."""
     repr_got = repr(chain)
-    repr_ext = "Chain: " + chain._link_symbol.join(chain.path)
+    repr_ext = "Chain: " + chain._link_symbol.join(
+        [full_to_camel(t) for t in chain.path]
+    )
     assert repr_got == repr_ext, "Unexpected repr of TableChain object."
